+++ ChaoticComets/Assets/Scripts/Saving Scripts/BetweenScenes.cs
@@ -2,10 +2,6 @@
     public static int PlayerCount = 1; // Defaut value of 1 player for debugging. Alternative is '2'.
     public static int Difficulty = 1; // Default value of Normal difficulty for debugging. Alternatives are '0' Easy, and '2' Hard.
 
-    // Default value of volume is full volume. Once Chaotic Comets is run once on a system, PlayerPrefs overwrites this value upon startup.
-    public static float MusicVolume = 1.0f;
-    public static float SFXVolume = 1.0f;
-
     // By default, the game assumes it's not being resumed from a save. Until it either...
     // A: New game is started, and first level is finished
     // B: Game is resumed from a found save file
@@ -23,10 +19,12 @@
     // This is because iterating on floats causes counting errors eventually with any programming language.
     // 10 = 1.0, or the base of upgrades. They'll be iterated by 1 (converted to 0.1 float) each time an upgrade is performed.
     // In order, upgrades are: Top speed, braking efficiency, fire rate, shot speed.
-    public static int[] UpgradesP1 = { 10, 10, 10, 10 };
-    public static int[] UpgradesP2 = { 10, 10, 10, 10 };
+    public static int[][] PlayerShopUpgrades = new int[2][] {
+        new int[] { 10, 10, 10, 10 },
+        new int[] { 10, 10, 10, 10 }
+    };
 
     // Credits and life counter are tracked here, because after a store is over, saves are erased but these two stats are still required
-    public static int[] playerShopCredits = { 0, 0 };
-    public static int[] playerShopLives = { 0, 0 };
+    public static int[] PlayerShopCredits = { 0, 0 };
+    public static int[] PlayerShopLives = { 0, 0 };
 }