+++ ChaoticComets/Assets/Scripts/Shop Scripts/ShopScript.ButtonActions.cs
@@ -7,10 +7,12 @@ using UnityEngine.UI;
 public partial class ShopScript : MonoBehaviour
 {
 
-    // Code for shop's Ready system. If both players are ready, shop closes.
+
+    /* ------------------------------------------------------------------------------------------------------------------
+    * Ready System - If both players are ready, shop closes.
+    * ------------------------------------------------------------------------------------------------------------------ */
     public void ReadyUp(int plrReady)
     {
-        plrReady -= 1; // Make compatible with array
         Navigation customNav = new Navigation { mode = Navigation.Mode.Explicit };
         Button[] listOfButtons = FindObjectsOfType<Button>();
 
@@ -24,13 +26,13 @@ public partial class ShopScript : MonoBehaviour
         string btnStartToCheck = "P" + (plrReady + 1);
         foreach (Button btn in listOfButtons)
         {
-            print(btn.transform.name);
             if (btn.transform.name.StartsWith(btnStartToCheck) && !btn.transform.name.EndsWith("Ready"))
             {
                 btn.GetComponent<Button>().interactable = !plrIsReady[plrReady];
             }
         }
 
+        // If all players are ready, start the next level
         if (plrIsReady.All(x => x))
         {
             for (int i = 0; i < plrIsReady.Length; i++)
@@ -43,80 +45,55 @@ public partial class ShopScript : MonoBehaviour
     }
 
     /* ------------------------------------------------------------------------------------------------------------------
-    * Upgrades code.
-    * They must be separate functions due to how Unity handles button press scripting. Only one variable can be passed per button.
+    * Upgrade Button - A button can only pass one variable. Passes which player pressed it, and figures out which one was pressed in code.
     * ------------------------------------------------------------------------------------------------------------------ */
-
-    public void PerformUpgradeP1(int whichUpgrade)
+    public void PerformUpgrade(int player)
     {
-        Button buttonUpgrading = ShopRefs.plrEventSystems[0].currentSelectedGameObject.GetComponent<Button>();
-        if (buttonUpgrading.name.StartsWith("P1"))
-        {
-            if (BetweenScenes.UpgradesP1[whichUpgrade] < upgradeCap)
-            {
+        // Find button pressed, turn the last character in the name into an integer
+        Button buttonUpgrading = ShopRefs.plrEventSystems[player].currentSelectedGameObject.GetComponent<Button>();
+        int whichUpgrade = int.Parse(buttonUpgrading.name.Last().ToString());
 
-                int price = baseUpgradePrice + priceIncreasePerLevel * (BetweenScenes.UpgradesP1[whichUpgrade] - 10);
-                if (BetweenScenes.playerShopCredits[0] >= price)
-                {
-                    buttonUpgrading.GetComponent<AudioSource>().Play();
-                    BetweenScenes.UpgradesP1[whichUpgrade] += 1;
-                    BetweenScenes.playerShopCredits[0] -= price;
-                    print($"Upgrades: {string.Join(",", BetweenScenes.UpgradesP1)} Credits left: {BetweenScenes.playerShopCredits[0]}");
-                }
-                else
-                {
-                    StartCoroutine(FlashCreditsRed(0));
-                }
+        // If the upgrade hasn't reached the upgrade cap, then attempt the upgrade
+        // If it can be afforded, perform it. If not, flash the credits text red
+        if (BetweenScenes.PlayerShopUpgrades[player][whichUpgrade] < upgradeCap)
+        {
+            int price = baseUpgradePrice + priceIncreasePerLevel * (BetweenScenes.PlayerShopUpgrades[player][whichUpgrade] - 10);
+            if (BetweenScenes.PlayerShopCredits[player] >= price)
+            {
+                buttonUpgrading.GetComponent<AudioSource>().Play();
+                BetweenScenes.PlayerShopUpgrades[player][whichUpgrade] += 1;
+                BetweenScenes.PlayerShopCredits[player] -= price;
+                print($"Upgrades: {string.Join(",", BetweenScenes.PlayerShopUpgrades[player])} Credits left: {BetweenScenes.PlayerShopCredits[player]}");
+                UpdateButtonText();
+            }
+            else
+            {
+                StartCoroutine(FlashCreditsRed(player));
             }
         }
-        UpdateButtonText();
     }
 
-    public void PerformUpgradeP2(int whichUpgrade)
-    {
-        Button buttonUpgrading = ShopRefs.plrEventSystems[1].currentSelectedGameObject.GetComponent<Button>();
-        if (buttonUpgrading.name.StartsWith("P2"))
-        {
-            if (BetweenScenes.UpgradesP2[whichUpgrade] < upgradeCap)
-            {
-                int price = baseUpgradePrice + priceIncreasePerLevel * (BetweenScenes.UpgradesP2[whichUpgrade] - 10);
-                if (BetweenScenes.playerShopCredits[1] >= price)
-                {
-                    buttonUpgrading.GetComponent<AudioSource>().Play();
-                    BetweenScenes.UpgradesP2[whichUpgrade] += 1;
-                    BetweenScenes.playerShopCredits[1] -= price;
-                    print($"Upgrades: {string.Join(",", BetweenScenes.UpgradesP2)} Credits left: {BetweenScenes.playerShopCredits[1] }");
-                }
-                else
-                {
-                    StartCoroutine(FlashCreditsRed(1));
-                }
-            }
-        }
-        UpdateButtonText();
-    }
-
-    // Give a life to player number 'i'
+    /* ------------------------------------------------------------------------------------------------------------------
+    * Life Transfer Button - Allows player to transfer a life to another player.
+    * ------------------------------------------------------------------------------------------------------------------ */
     public void GiveLife(int plrSendingLife)
     {
-        plrSendingLife -= 1; // Make compatible with array
-        int plrReceivingLife;
-
         // TODO support more than two players. Probably automatically determine the player with the least lives.
+        int plrReceivingLife;
         if (plrSendingLife == 0) { plrReceivingLife = 1; }
         else /*(plrSendingLife == 1)*/ { plrReceivingLife = 0; }
 
         // If player can send a life, then give a life
-        if (BetweenScenes.playerShopLives[plrSendingLife] > 1 && BetweenScenes.playerShopCredits[plrSendingLife] >= 500)
+        if (BetweenScenes.PlayerShopLives[plrSendingLife] > 1 && BetweenScenes.PlayerShopCredits[plrSendingLife] >= 500)
         {
-            Button buttonGivingLife = ShopRefs.plrEventSystems[0].currentSelectedGameObject.GetComponent<Button>();
+            Button buttonGivingLife = ShopRefs.plrEventSystems[plrSendingLife].currentSelectedGameObject.GetComponent<Button>();
             buttonGivingLife.GetComponent<AudioSource>().Play();
 
-            BetweenScenes.playerShopCredits[plrSendingLife] -= 500;
-            BetweenScenes.playerShopLives[plrSendingLife] -= 1;
-            BetweenScenes.playerShopLives[plrReceivingLife] += 1;
+            BetweenScenes.PlayerShopCredits[plrSendingLife] -= 500;
+            BetweenScenes.PlayerShopLives[plrSendingLife] -= 1;
+            BetweenScenes.PlayerShopLives[plrReceivingLife] += 1;
             // Reset dead player's shields to full when receiving their first new life
-            if (BetweenScenes.playerShopLives[plrReceivingLife] == 1)
+            if (BetweenScenes.PlayerShopLives[plrReceivingLife] == 1)
             {
                 ShopRefs.listOfPlrShieldBars[plrReceivingLife].fillAmount = 1f;
             }
@@ -124,11 +101,11 @@ public partial class ShopScript : MonoBehaviour
         else
         {
             // If life transfer fails, flash one or both of the offending statistics
-            if (BetweenScenes.playerShopCredits[plrSendingLife] < 500)
+            if (BetweenScenes.PlayerShopCredits[plrSendingLife] < 500)
             {
                 StartCoroutine(FlashCreditsRed(plrSendingLife));
             }
-            if (BetweenScenes.playerShopLives[plrSendingLife] == 1)
+            if (BetweenScenes.PlayerShopLives[plrSendingLife] == 1)
             {
                 StartCoroutine(FlashLivesRed(plrSendingLife));
             }
@@ -136,6 +113,9 @@ public partial class ShopScript : MonoBehaviour
         UpdateButtonText();
     }
 
+    /* ------------------------------------------------------------------------------------------------------------------
+    * Stat Flashing Methods - Gives player feedback if not enough credits or lives to perform an action.
+    * ------------------------------------------------------------------------------------------------------------------ */
     private bool[] isAlreadyFlashingCredits = { false, false };
     private IEnumerator FlashCreditsRed(int playerFlashing)
     {
@@ -149,6 +129,7 @@ public partial class ShopScript : MonoBehaviour
             isAlreadyFlashingCredits[playerFlashing] = false;
         }
     }
+
     private bool[] isAlreadyFlashingLives = { false, false };
     private IEnumerator FlashLivesRed(int playerFlashing)
     {